applications/test2/requirements.txt
flask==2.0.2
grpcio==1.41.1
grpcio-tools==1.41.1